High Precision and Code Cleanup
Added a 'High Precision' mode for adjusting the mirror distance in the menus. It is on the bottom right of page 1 for the QM. And in the extras menu of the Action Menu.
    The High Precision value is adjustable in Mod Settings.

Changed the MelonPref names of CanPickup*Mirror to all just be CanPickupMirror. This allowed me to cleanup some of Davi's code they made to handle that.